requirements.txt
numpy==1.24.1
opencv_python==4.7.0.68
pandas==1.5.2
pymediainfo==6.0.1
PyYAML==6.0
tqdm==4.64.1
ultralytics==8.0.9
comet-ml
fastapi[all]
pytest
modal-client